fix: moved file content from input.ts to inputs.ts
