+++ src/types/inputs.ts
@@ -1,3 +1,5 @@
+import { KeyboardTypeOptions, TextInput, TextInputProps, TextStyle } from 'react-native';
+
 export enum InputStatus {
   DEFAULT = 'default',
   ERROR = 'error',
@@ -18,6 +20,33 @@ export enum KeyboardTypes {
   URL = 'url',
   VISIBLE_PASSWORD = 'visible-password',
   WEB_SEARCH = 'web-search',
+}
+
+export interface InputProps extends Omit<TextInputProps, 'style | multiline'> {
+  disabled?: boolean;
+  endEnhancer?: React.ReactNode;
+  handleInputRef?: (ref: TextInput) => void;
+  keyboardType?: KeyboardTypeOptions;
+  label?: string;
+  message?: string;
+  startEnhancer?: React.ReactNode;
+  status?: InputStatus;
+  testId?: string;
+  textAlign?: Exclude<TextStyle['textAlign'], 'auto' | 'justify'>;
+}
+
+export interface PasswordInputProps extends InputProps {}
+
+export interface TextAreaInputProps extends Omit<TextInputProps, 'style | multiline'> {
+  disabled?: boolean;
+  endEnhancer?: React.ReactNode;
+  handleInputRef?: (ref: TextInput) => void;
+  keyboardType?: KeyboardTypeOptions;
+  label?: string;
+  numberOfLines?: number;
+  startEnhancer?: React.ReactNode;
+  testId?: string;
+  textAlign?: Exclude<TextStyle['textAlign'], 'auto' | 'justify'>;
 }
 
 export interface WebsiteUrlInputProps {
